Add a sidebar of rooms
Adding the sidebar with a list of rooms involves doing the following:

Changing the mount function to show a list of rooms.

Adding the sidebar to the render function.

Adding a room_link private function.

Problem: the room appears below the sidebar instead of to the right of the sidebar.