Change the selected room
Change the selected room when making a selection in the sidebar. This involves the following changes:

Change the link destination in the room_link function.

Change the mount function to fetch the desired room id.

I still don't understand why the selected room is below the sidebar instead of next to the sidebar.